Fix possible deadlock scenario
There is a scenario where successive incoming websocket messages
(e.g. 2 or more in a microsecond) could cause a possible deadlock.
The deadlock occurs when the App receiving messages on the incoming
channel tries to send an outgoing message but blocks as the gooey
loop is stuck on waiting to send the next incoming message.  So the
order goes like this:  message 1 from the WS is sent for handling,
message 2 arrives but blocks when trying to send the message on the
incoming channel as the App is processing message 1, and then the
App deadlocks on the outgoing channel if it is sending a response as
the gooey loop can't address the outgoing message since it's blocked
on the incoming channel.

Here the simple fix is to remove the daisy chain of reading a message
from the websocket and forwarding it to the gooey loop.  This was
set up this way because originally would attempt to unmarshal the
message to check for any special tags addressed to it.  Of course,
this scheme was removed but the daisy chain was left in place.
Removing the daisy chain works as now the reading of the websocket,
which is run in its own goroutine is blocked until the App can
acknowledge it.